+++ lib/djv/App/ColorTool.cpp
@@ -44,7 +44,6 @@ namespace djv
             auto levelsWidget = ui::LevelsWidget::create(context, app->getSettings(), viewportModel);
             auto exposureWidget = ui::ExposureWidget::create(context, viewportModel);
             auto softClipWidget = ui::SoftClipWidget::create(context, viewportModel);
-            auto clipWarningWidget = ui::ClipWarningWidget::create(context, viewportModel);
 
             auto layout = ftk::VerticalLayout::create(context);
             layout->setSpacingRole(ftk::SizeRole::Border);
@@ -68,9 +67,6 @@ namespace djv
             p.bellows["SoftClip"] = ftk::Bellows::create(context, "Soft Clip", layout);
             p.bellows["SoftClip"]->setWidget(softClipWidget);
             p.bellows["SoftClip"]->setToolWidget(softClipWidget->getEnabledCheckBox());
-            p.bellows["ClipWarning"] = ftk::Bellows::create(context, "Clip Warning", layout);
-            p.bellows["ClipWarning"]->setWidget(clipWarningWidget);
-            p.bellows["ClipWarning"]->setToolWidget(clipWarningWidget->getEnabledCheckBox());
             _setWidget(layout);
 
             _loadSettings(p.bellows);